+++ changelog.md
@@ -5,3 +5,4 @@ This release has a bug fix.
 ### New Features
 
 -   Fixed bug with presets disappearing when selecting certain miners
+-   Added clock and voltage tuning capabilities for Bitmain S19Js in the tune tab